# mjfast-admin/src/main/resources/public/swagger/index.yaml
swagger: '2.0'
info:
  description: 是一套轻量级权限管理系统，极低门槛，拿来即用。设计之初，就非常注重安全性，为企业系统保驾护航，让一切都变得如此简单。
  version: 3.0.0
  title: 企业快速开发框架在线API
basePath: /admin
schemes:
- http
paths:
...
  /sys/user/list:
    get:
      tags:
      - 管理员管理
      summary: 管理员列表
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: order
        description: 排序方式，如：asc、desc
        in: query
        type: string
      - name: username
        description: 用户名
        in: query
        type: string
      responses:
        '200':
          description: 返回管理员列表
          schema:
            $ref: '#/definitions/SysUserEntity'
...
  /sys/user/info/{userId}:
    get:
      tags:
      - 管理员管理
      summary: 获取管理员信息
      produces:
      - application/json
      parameters:
      - name: userId
        description: 用户ID
        in: path
        type: integer
        required: true
      responses:
        '200':
          description: 返回管理员信息
          schema:
            $ref: '#/definitions/SysUserEntity'
  /sys/user/password:
    post:
      tags:
      - 管理员管理
      summary: 修改密码
      produces:
      - application/json
      parameters:
      - name: password
        description: 原密码
        in: query
        type: string
        required: true
      - name: newPassword
        description: 新密码
        in: query
        type: string
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/save:
    post:
      tags:
      - 管理员管理
      summary: 添加管理员
      produces:
      - application/json
      parameters:
      - name: body
        description: 管理员对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysUserEntityEdit'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/update:
    post:
      tags:
      - 管理员管理
      summary: 修改管理员
      produces:
      - application/json
      parameters:
      - name: body
        description: 管理员对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysUserEntityEdit'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/delete:
    post:
      tags:
      - 管理员管理
      summary: 删除管理员
      produces:
      - application/json
      parameters:
      - name: body
        description: 用户ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/menu/info/{menuId}:
    get:
      tags:
      - 菜单管理
      summary: 获取菜单信息
      produces:
      - application/json
      parameters:
      - name: menuId
        description: 菜单ID
        in: path
        type: integer
        required: true
      responses:
        '200':
          description: 返回菜单信息
          schema:
            $ref: '#/definitions/SysRoleEntity'
  /sys/menu/save:
    post:
      tags:
      - 菜单管理
      summary: 添加菜单
      produces:
      - application/json
      parameters:
      - name: body
        description: 菜单对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysMenuEntityEdit'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/update:
    post:
      tags:
      - 菜单管理
      summary: 修改菜单
      produces:
      - application/json
      parameters:
      - name: body
        description: 菜单对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysMenuEntityEdit'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/delete:
    post:
      tags:
      - 菜单管理
      summary: 删除菜单
      produces:
      - application/json
      parameters:
      - name: body
        description: 菜单ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
  SysUserEntity:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      password:
        description: 密码
        type: string
      email:
        description: 邮箱
        type: string
      mobile:
        description: 手机号
        type: string
      status:
        description: 状态  0：禁用   1：正常
        type: integer
        format: int32
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 创建者ID
        type: integer
        format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
  SysUserEntityEdit:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      password:
        description: 密码
        type: string
      email:
        description: 邮箱
        type: string
      mobile:
        description: 手机号
        type: string
      status:
        description: 状态  0：禁用   1：正常
        type: integer
        format: int32
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int32
...
  SysRoleEntity:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名称
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 创建者ID
        type: integer
        format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
...
  SysMenuEntityEdit:
    type: object
    properties:
      menuId:
        description: 菜单ID
        type: integer
        format: int64
      name:
        description: 菜单名称
        type: string
      parentId:
        description: 父菜单ID，一级菜单为0
        type: integer
        format: int64
      url:
        description: 菜单URL
        type: string
      perms:
        description: 授权标识
        type: string
      type:
        description: 类型     0：目录   1：菜单   2：按钮
        type: integer
        format: int32
      icon:
        description: 菜单图标
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
...